+++ src/Polly.SharedSpecs/Helpers/ContextualPolicyTResultExtensionsAsync.cs
@@ -10,14 +10,14 @@ namespace Polly.Specs.Helpers
     public static class ContextualPolicyTResultExtensionsAsync
     {
 
-        public static Task<TResult> RaiseResultSequenceAsync<TResult>(this ContextualPolicy<TResult> policy,
+        public static Task<TResult> RaiseResultSequenceAsync<TResult>(this Policy<TResult> policy,
     IDictionary<string, object> contextData,
     params TResult[] resultsToRaise)
         {
             return policy.RaiseResultSequenceAsync(contextData, CancellationToken.None, resultsToRaise.ToList());
         }
 
-        public static Task<TResult> RaiseResultSequenceAsync<TResult>(this ContextualPolicy<TResult> policy, IDictionary<string, object> contextData, CancellationToken cancellationToken, IEnumerable<TResult> resultsToRaise)
+        public static Task<TResult> RaiseResultSequenceAsync<TResult>(this Policy<TResult> policy, IDictionary<string, object> contextData, CancellationToken cancellationToken, IEnumerable<TResult> resultsToRaise)
         {
             var enumerator = resultsToRaise.GetEnumerator();
 
@@ -32,12 +32,12 @@ namespace Polly.Specs.Helpers
             }, contextData, cancellationToken);
         }
 
-        public static Task<PolicyResult<TResult>> RaiseResultSequenceOnExecuteAndCaptureAsync<TResult>(this ContextualPolicy<TResult> policy, IDictionary<string, object> contextData, params TResult[] resultsToRaise)
+        public static Task<PolicyResult<TResult>> RaiseResultSequenceOnExecuteAndCaptureAsync<TResult>(this Policy<TResult> policy, IDictionary<string, object> contextData, params TResult[] resultsToRaise)
         {
             return policy.RaiseResultSequenceOnExecuteAndCaptureAsync(contextData, resultsToRaise.ToList());
         }
 
-        public static Task<PolicyResult<TResult>> RaiseResultSequenceOnExecuteAndCaptureAsync<TResult>(this ContextualPolicy<TResult> policy, IDictionary<string, object> contextData, IEnumerable<TResult> resultsToRaise)
+        public static Task<PolicyResult<TResult>> RaiseResultSequenceOnExecuteAndCaptureAsync<TResult>(this Policy<TResult> policy, IDictionary<string, object> contextData, IEnumerable<TResult> resultsToRaise)
         {
             var enumerator = resultsToRaise.GetEnumerator();